Finished main logic of twitch webhook handler
We have to handle webhook logic in a separate Flask backend from the
actual bot code. Both will need to be integrated, likely using RabbitMQ
to communicate between the two services, which should be an architecture
that is deployable in production. Message verification is done for the
webhook side and actual notification handling (send message with
RabbitMQ to bot code) needs to be implemented.