Add server join config for bot (not quite working)
When joining server bot sends DM for button to activate a modal
for user to select channel for bot to send @ notifs in for streamsnipe
except the sending of the modal when clicking the button errors.

Also, the channel select dropdown is empty since it is in DMs, so find
a way to have it be in the server but hidden from everyone except server
owner somehow?