Refine DB models/schema
Added relationships and cascade from guild to user subscriptions.
Cascade for user subscriptions to streamer table not possible since the
cascade depends conditionally on user subscriptions where all user subs
for a streamer need to be gone whereas with guild to user subs if we
delete guild we just delete user subs associated with the guild on no
condition, so a trigger or custom cascade might be necessary.