Fix foreign key issue for user sub table
Bruh moment. We have foreign key on user sub table for streamer but we
don't have a cascade/auto-insert set up for the streamer table so the
foreign keys for user sub entries refer to nothing. Although triggers
would be ideal, SQLAlchemy isn't really set up to manage DB stuff at
that low level so we manually check if streamer is seen before in
streamer table and insert it if it doesn't exist yet before inserting
user sub so the foreign key constraint is met.