Added custom cascade for user sub table to streamer table upon deletion
Decided to do manual queries rather than event handling or trigger.
A lot simpler to manage even if it is more verbose.
We also have to account for cascade of guild to user sub when bot
is removed from a guild since it might delete some user sub entries
which means that the streamer table might need to be cleaned up as well.
We use left outerjoin to match streamer IDs from user sub table against
the streamer table. Any streamer that no longer has any references in
the user sub table would result in a null on the user sub side when
joined with the streamer table (which is on the left) so we can check
for the nulls to figure out what streamer rows to purge.
Tested the code and it seems to be working except for one time where it
timed out? Could be a persistent issue that requires more test runs
to figure out if it is consistent or not.