Add unit tests and revamed file structure for tests
Still mostly boilerplate to test out pytest and writing unit tests for
the bot. Will need to flesh out the unit tests for all functions in the
bot and then start integration tests
which are close to end to end tests that use a testing channel in a live
discord dev server.

Subdirectories are packages (have dunder init files that are empty) for
easy importing. Also pytest importing the bot file caused infinite loop
basically (imported files are executed so it ran the bot...)
Fixed by adding the name main idiom check so the bot only runs
when executed as a script.