Add parse streamer unit tests and more error handling
Parse streamer url error handling added. It used to accept invalid urls
that had a valid streamer name in the path at the end
so technically it shouldn't work but the assumption was
that the streamer name was at the end and it just needed to look like a
URL but we didn't check the URL itself. Now we do check for a twitch.tv
URL with regex