Add check for dupe streamers in parse streamers from command function
Also painstakingly updated unit tests since we convert back and forth
between sets and lists, and lists from sets don't preserve order
so old asserts are lists with specific
ordering of elements which don't work.

Check arguments separately so we can assert on them and sort the result
to get a consistent output to assert against.