Fix hash error
Redudant API call fix implementation used dicts which are unhashable
when put into sets to get rid of dupes.
Created a custom frozen dataclass with hash method to allow for
hashing to store api response and make things more self documenting
compared to just a tuple with indices when used to access the data.